+++ tests/header_contracts/error_header.cpp
@@ -1,0 +1,4 @@
+#include <vosp/contracts/error.hpp>
+
+static_assert(!vosp::contracts::Error<int>);
+